+++ common/repositories/pack_price_limits.py
@@ -1,3 +1,5 @@
+from decimal import Decimal
+
 from redis.asyncio import Redis
 
 from common.catalog.packages import PACKAGE_SIZES, PACKAGE_UNIT_COUNT
@@ -6,28 +8,28 @@ _KEY = "settings:pack_price_limit"
 _DEFAULT_UNIT_PRICE_CAP = 1000
 
 
-def _default_limit(size: int) -> int:
-    return _DEFAULT_UNIT_PRICE_CAP * PACKAGE_UNIT_COUNT[size]
+def _default_limit(size: int) -> Decimal:
+    return Decimal(_DEFAULT_UNIT_PRICE_CAP * PACKAGE_UNIT_COUNT[size])
 
 
 class PackPriceLimitRepository:
     def __init__(self, *, redis: Redis) -> None:
         self._redis = redis
 
-    async def get(self, *, size: int) -> int:
+    async def get(self, *, size: int) -> Decimal:
         value = await self._redis.hget(_KEY, str(size))
         if value is None:
             return _default_limit(size)
-        return int(value)
+        return Decimal(value)
 
-    async def get_all(self) -> dict[int, int]:
+    async def get_all(self) -> dict[int, Decimal]:
         raw = await self._redis.hgetall(_KEY)
         return {
-            size: int(raw[str(size)]) if str(size) in raw else _default_limit(size)
+            size: Decimal(raw[str(size)]) if str(size) in raw else _default_limit(size)
             for size in PACKAGE_SIZES
         }
 
-    async def set(self, *, size: int, limit: int) -> None:
+    async def set(self, *, size: int, limit: Decimal) -> None:
         await self._redis.hset(_KEY, str(size), str(limit))
 
     async def reset(self, *, size: int) -> None:
